Update index.js
When we use option destRowTemplate as function, we get an array. And when we use "join('')" on array we get a string whisout separator.
If in formatter function we return some string, out variable fileList be like that ["string 1", "string 2", ...] and when we put contents to file, we get somethink like that "string 1string 2".
If in formatter function we return some json object, we cant put correct json to file, because we use join('') on array when destRowTemplate is set.
I think no need use join ever

Sorry for my bad english